% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenodata.R
\name{create_phenodata}
\alias{create_phenodata}
\title{Create the Phenodata for affy input}
\usage{
create_phenodata(experiment.set, data.path)
}
\arguments{
\item{experiment.set}{vector, store the experiment set in vector}

\item{data.path}{str, the input raw cell directory path}
}
\value{
this function will return a tbl_df data , if the filename not exist in current directory, show error message
}
\description{
This function will create the phenodata which used in loading cel files
}
\examples{
create_phenodata(experiment.set=c("control_1","control_2","case_1","case_2"), data.path='./cel_data_dir')
}

