% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualization fo the accident happening in certain state during a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a integer, the  code number of state}

\item{year}{a integer, indicated the year's  of the traffic accidents}
}
\value{
a graphical object, this function will return a plot with the accident happening at certain state with certain year,
if the STATE number is not exist, show error message, if the choosen state have no accident happended, show errors
}
\description{
This function will visualize the traffic accidents happend during certain year  within certain state into the geographic view
}
\examples{
\dontrun{
fars_map_state(state.num = 1, year = 2014)
}
}
