% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Load the csv data into R as tbl_df}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{str, the input file's name}
}
\value{
this function will return a tbl_df data , if the filename not exist in current directory, show error message
}
\description{
This function will use the readr::read_csv to load the .csv file into R
and turn the data into tbl_df data structure
}
\examples{
\dontrun{
fars_read(filename="./data/accident_2003.csv.bz2")
}
}

