% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get all accidents happend according to months within certain year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the years want to query the accident , and can be a vector}
}
\value{
a list with data.frame elements, this function will return the tbl_df  with month and year column within certain years
if the input years didnt contain in the dataset, show error message.
}
\description{
This function will first generate the files name according to the input year, then loadding the generated filename,
select the month and years column to output
}
\examples{
\dontun{
fars_read_years(years=2013)
}
}

