% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the accidence number happended according to months within certain year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vectors contained integer indicated the years we want to query the accidennt happend}
}
\value{
a data.frame, according to the years with the information of the accidents
}
\description{
This function will first load the data accoding tot years input
}
\examples{
\dontrun{
fars_summarize_years(years = c(2013,2014))
}
}
