% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Generate the file name with the given the request year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a integer or string, represented the given year}
}
\value{
this function will return the file name according to the input year
}
\description{
This function will generate the file name according to the given years and will be used by another function
}
\examples{
\dontrun{make_filename(year=2012)
}
}
