% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{create_phenodata}
\alias{create_phenodata}
\title{Create the Phenodata for affy input}
\usage{
create_phenodata(experiment.set)
}
\arguments{
\item{experiment.set}{vector, store the experiment set in vector}
}
\value{
this function will return a tbl_df data , if the filename not exist in current directory, show error message
}
\description{
This function will create the phenodata which used in loading cel files
}
\examples{
create_phenodata(experiment.set=c("control_1","control_2","case_1","case_2"))
}

