% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualization fo the accident happening in certain state during a specific year}
\usage{
fars_map_state(state.num, year)
}
\description{
@description  This function will visualize the traffic accidents happend during certain year  within certain state into the geographic view
 @param state.num a integer, the  code number of state
 @param year a integer, indicated the year's  of the traffic accidents
 @return a graphical object, this function will return a plot with the accident happening at certain state with certain year,
 if the STATE number is not exist, show error message, if the choosen state have no accident happended, show errors
 @examples  fars_map_state(state.num = 1, year = 2014)
 @importFrom  dplyr::filter
 @importFrom maps::map
 @importFrom graphics::points
}

